+++ src/app/item/table/enhancedTableHead/EnhancedTableHead.tsx
@@ -5,9 +5,9 @@ import { visuallyHidden } from '@mui/utils';
 import {
   Box, Checkbox, TableCell, TableRow,
 } from '@mui/material';
-import { Collection } from '../../../../shared/models/items/itemsCollection';
+import { Collection } from '../../../../shared/models/items/itemsCollection.model';
 import getHeadCell from '../../../../shared/utils/getHeadCell';
-import HeadCell from '../../../../shared/models/itemsTable/headCell';
+import HeadCell from '../../../../shared/models/itemsTable/headCell.model';
 import { useAppSelector } from '../../../../shared/hooks/hooks';
 import styles from './styles.module.scss';
 import Order from '../../../../shared/models/items/order.type';
