+++ src/shared/models/allCollections/allCollectionsResponse.model.ts
