+++ src/shared/models/authState.model.ts
@@ -1,5 +1,5 @@
 // import { AuthResponse } from './authResponse.model';
-import InitialState from './state/initialState';
+import InitialState from './state/initialState.model';
 
 interface AuthState extends InitialState {
   userId: string,
