+++ src/shared/models/deleteCollection/deleteData.model.ts
