+++ src/shared/models/items/itemStructure.model.ts
