+++ src/shared/models/items/newItem.type.ts
