+++ src/shared/models/items/newItemRequest.model.ts
