+++ src/shared/models/state/allCollection.module.ts
@@ -1,5 +1,5 @@
-import InitialState from './initialState';
-import AllCollectionsResponse from '../allCollections/allCollectionsResponse';
+import InitialState from './initialState.model';
+import AllCollectionsResponse from '../allCollections/allCollectionsResponse.model';
 
 interface AllCollection extends InitialState {
   allCollections?: AllCollectionsResponse[];
