+++ src/shared/models/state/itemState.model.ts
@@ -1,5 +1,5 @@
-import ItemStructure from '../items/itemStructure';
-import InitialState from './initialState';
+import ItemStructure from '../items/itemStructure.model';
+import InitialState from './initialState.model';
 
 interface ItemState extends InitialState {
   item: ItemStructure | null,
