+++ src/shared/utils/convertItemsForTable.ts
@@ -1,4 +1,4 @@
-import ItemStructure from '../models/items/itemStructure';
+import ItemStructure from '../models/items/itemStructure.model';
 
 const convertItemsForTable = (items: ItemStructure[]): Record<string, string>[] => items.map((i) => {
   let y = { id: i.id, title: i.title };
