+++ src/shared/utils/getSelectedTitles.ts
@@ -1,4 +1,4 @@
-import ItemStructure from '../models/items/itemStructure';
+import ItemStructure from '../models/items/itemStructure.model';
 
 const getSelectedTitles = (title: ItemStructure[], selected: string[]): string[] => {
   const n = title.filter((i) => selected.includes(i.id));
